requirements.txt
alembic==1.7.6
anyio==3.5.0
apispec==3.3.2
argcomplete==1.12.3
asn1crypto==1.5.1
attrs==23.2.0
Babel==2.9.1
beautifulsoup4==4.10.0
blinker==1.4
boto3==1.34.32
botocore==1.34.32
cachelib==0.6.0
cachetools==4.2.4
cattrs==1.1.0
certifi==2023.11.17
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==5.0.1
commonmark==0.9.1
configparser==5.0.2
croniter==1.0.15
cryptography==42.0.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.4
dnspython==2.2.1
docutils==0.16
email-validator==1.1.3
Flask==1.1.4
Flask-AppBuilder==3.4.4
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
future==0.18.2
gitdb==4.0.9
GitPython==3.1.20
google-api-core==2.8.0
google-api-python-client==2.48.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.1
graphviz==0.19.1
greenlet==3.0.3
gunicorn==20.1.0
h11==0.12.0
hmsclient==0.1.1
httpcore==0.14.7
httplib2==0.20.4
httpx==0.22.0
idna==3.6
importlib-metadata==4.8.3
inflection==0.5.1
iso8601==1.0.2
isodate==0.6.1
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.1.0
JSON-log-formatter==0.4.0
jsonpath-ng==1.5.3
jsonschema==4.17.3
jsonschema-specifications==2023.12.1
lazy-object-proxy==1.7.1
lockfile==0.12.2
lxml==5.1.0
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==1.1.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mysql-client==0.0.1
mysql-connector==2.2.9
numpy==1.24.0
oauth2client==4.1.3
oauthlib==3.2.0
packaging==21.3
pandas==2.2.0
pendulum==2.1.2
ply==3.11
prison==0.2.1
protobuf==4.21.0
psutil==5.9.0
psycopg2-binary==2.9.9
py4j==0.10.9.3
pyairtable==1.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyboto3==1.13.18
pycparser==2.21
Pygments==2.11.2
PyJWT==1.7.1
PyMySQL==1.0.2
pyOpenSSL==24.0.0
pyparsing==3.0.7
pyrsistent==0.14.0
python-daemon==2.3.0
python-dateutil==2.8.2
python-json-logger==2.0.2
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.3
pytzdata==2020.1
PyYAML==5.3.1
redshift-connector==2.0.905
referencing==0.33.0
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
retrying==1.3.3
rfc3986==1.5.0
rich==12.0.0
rpds-py==0.17.1
rsa==4.8
s3transfer==0.10.0
scramp==1.4.4
setproctitle==1.2.2
six==1.16.0
slugify==0.0.1
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.4.51
SQLAlchemy-JSONField==1.0.0
sqlalchemy-redshift==0.8.9
SQLAlchemy-Utils==0.38.2
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
thrift==0.15.0
tqdm==4.64.0
typing_extensions==4.2.0
tzdata==2023.4
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.18
watchtower==1.0.6
Werkzeug==1.0.1
wget==3.2
wrapt==1.14.0
WTForms==2.3.3
zipp==3.6.0
